# packages/contracts/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "MartHub API",
    "version": "1.0.0",
    "description": "Implemented foundation, authentication, account, and public catalog endpoints. Future inventory: docs/API.md."
  },
  "servers": [
    {
      "url": "/api/v1"
    }
  ],
  "paths": {
...
    "/products": {
      "get": {
        "operationId": "catalogListProducts",
        "description": "Search active, priced products by name, brand, or SKU; filter, sort, and paginate with deterministic tie breakers. Category includes direct children. All prices are VND integer strings.",
        "security": [],
        "responses": {
          "200": {
            "description": "Public product cards.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProductListResponse"
                }
              }
            }
          },
          "422": {
            "$ref": "#/components/responses/Error"
          },
          "500": {
            "$ref": "#/components/responses/Error"
          }
        },
        "parameters": [
          {
            "name": "q",
            "in": "query",
            "required": false,
            "description": "Normalized 2-80 character search text; all space-separated terms must match name, brand, or SKU.",
            "schema": {
              "type": "string",
              "minLength": 2,
              "maxLength": 80
            }
          },
          {
            "name": "category",
            "in": "query",
            "required": false,
            "description": "Active category slug; includes products in its direct children.",
            "schema": {
              "type": "string",
              "minLength": 1,
              "maxLength": 160,
              "pattern": "^[a-z0-9]+(-[a-z0-9]+)*$"
            }
          },
          {
            "name": "minPrice",
            "in": "query",
            "required": false,
            "description": "Inclusive current price in VND.",
            "schema": {
              "type": "string",
              "pattern": "^(0|[1-9][0-9]*)$",
              "maxLength": 19
            }
          },
          {
            "name": "maxPrice",
            "in": "query",
            "required": false,
            "description": "Inclusive current price in VND; must not be less than minPrice.",
            "schema": {
              "type": "string",
              "pattern": "^(0|[1-9][0-9]*)$",
              "maxLength": 19
            }
          },
          {
            "name": "availability",
            "in": "query",
            "required": false,
            "description": "Include only products with stock.",
            "schema": {
              "type": "string",
              "enum": ["in-stock"]
            }
          },
          {
            "name": "featured",
            "in": "query",
            "required": false,
            "description": "Match the curated featured flag.",
            "schema": {
              "type": "string",
              "enum": ["true", "false"]
            }
          },
          {
            "name": "new",
            "in": "query",
            "required": false,
            "description": "Match the curated new flag.",
            "schema": {
              "type": "string",
              "enum": ["true", "false"]
            }
          },
          {
            "name": "popular",
            "in": "query",
            "required": false,
            "description": "Match the curated popular flag.",
            "schema": {
              "type": "string",
              "enum": ["true", "false"]
            }
          },
          {
            "name": "sort",
            "in": "query",
            "required": false,
            "description": "Defaults to relevance when q is present, otherwise newest. Relevance requires q.",
            "schema": {
              "type": "string",
              "enum": [
                "relevance",
                "newest",
                "price-asc",
                "price-desc",
                "popular"
              ]
            }
          },
          {
            "name": "page",
            "in": "query",
            "required": false,
            "description": "One-based page, maximum 1000; default 1.",
            "schema": {
              "type": "string",
              "pattern": "^[1-9][0-9]*$",
              "default": "1"
            }
          },
          {
            "name": "perPage",
            "in": "query",
            "required": false,
            "description": "Items per page, maximum 60; default 24.",
            "schema": {
              "type": "string",
              "pattern": "^[1-9][0-9]*$",
              "default": "24"
            }
          }
        ]
      }
    },
    "/products/{slug}": {
      "get": {
        "operationId": "catalogGetProduct",
        "description": "Return active product detail with current VND price, ordered images, and stock availability.",
        "security": [],
        "parameters": [
          {
            "name": "slug",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 1,
              "maxLength": 160,
              "pattern": "^[a-z0-9]+(-[a-z0-9]+)*$"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Public product detail.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProductResponse"
                }
              }
            }
          },
          "404": {
            "$ref": "#/components/responses/Error"
          },
          "422": {
            "$ref": "#/components/responses/Error"
          },
          "500": {
            "$ref": "#/components/responses/Error"
          }
        }
      }
    },
...
    "/admin": {
      "get": {
        "operationId": "adminGetSession",
        "description": "Authorize the current database-backed Admin session and return only its safe identity projection.",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Current authorized Admin identity.",
            "headers": {
              "Cache-Control": {
                "schema": {
                  "type": "string",
                  "const": "no-store"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminSessionResponse"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/AuthError"
          },
          "403": {
            "$ref": "#/components/responses/AuthError"
          },
          "500": {
            "$ref": "#/components/responses/AuthError"
          }
        }
      }
    },
...
    "/admin/products": {
      "get": {
        "operationId": "adminProductsList",
        "description": "Search, filter, and paginate draft, active, and archived products for administration.",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "q",
            "in": "query",
            "schema": {
              "type": "string",
              "minLength": 1,
              "maxLength": 80
            }
          },
          {
            "name": "status",
            "in": "query",
            "schema": {
              "type": "string",
              "enum": ["DRAFT", "ACTIVE", "ARCHIVED"]
            }
          },
          {
            "name": "categoryId",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "page",
            "in": "query",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 1000,
              "default": 1
            }
          },
          {
            "name": "perPage",
            "in": "query",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 60,
              "default": 24
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Administrative product list.",
            "headers": {
              "Cache-Control": {
                "schema": {
                  "type": "string",
                  "const": "no-store"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminProductListResponse"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/AuthError"
          },
          "403": {
            "$ref": "#/components/responses/AuthError"
          },
          "422": {
            "$ref": "#/components/responses/AuthError"
          },
          "500": {
            "$ref": "#/components/responses/AuthError"
          }
        }
      },
      "post": {
        "operationId": "adminProductsCreate",
        "description": "Create a draft product with an initial exact VND price and zero stock.",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AdminProductCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Draft product created.",
            "headers": {
              "Cache-Control": {
                "schema": {
                  "type": "string",
                  "const": "no-store"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminProductResponse"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/AuthError"
          },
          "401": {
            "$ref": "#/components/responses/AuthError"
          },
          "403": {
            "$ref": "#/components/responses/AuthError"
          },
          "404": {
            "$ref": "#/components/responses/AuthError"
          },
          "409": {
            "$ref": "#/components/responses/AuthError"
          },
          "413": {
            "$ref": "#/components/responses/AuthError"
          },
          "415": {
            "$ref": "#/components/responses/AuthError"
          },
          "422": {
            "$ref": "#/components/responses/AuthError"
          },
          "500": {
            "$ref": "#/components/responses/AuthError"
          }
        }
      }
    },
    "/admin/products/{productId}": {
      "parameters": [
        {
          "$ref": "#/components/parameters/AdminProductId"
        }
      ],
      "get": {
        "operationId": "adminProductsGet",
        "description": "Retrieve complete product administration detail.",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Administrative product detail.",
            "headers": {
              "Cache-Control": {
                "schema": {
                  "type": "string",
                  "const": "no-store"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminProductResponse"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/AuthError"
          },
          "403": {
            "$ref": "#/components/responses/AuthError"
          },
          "404": {
            "$ref": "#/components/responses/AuthError"
          },
          "500": {
            "$ref": "#/components/responses/AuthError"
          }
        }
      },
      "patch": {
        "operationId": "adminProductsUpdate",
        "description": "Update allowed catalog fields. SKU, slug, price, inventory, and status use immutable or dedicated commands.",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AdminProductUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Product updated.",
            "headers": {
              "Cache-Control": {
                "schema": {
                  "type": "string",
                  "const": "no-store"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminProductResponse"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/AuthError"
          },
          "401": {
            "$ref": "#/components/responses/AuthError"
          },
          "403": {
            "$ref": "#/components/responses/AuthError"
          },
          "404": {
            "$ref": "#/components/responses/AuthError"
          },
          "409": {
            "$ref": "#/components/responses/AuthError"
          },
          "413": {
            "$ref": "#/components/responses/AuthError"
          },
          "415": {
            "$ref": "#/components/responses/AuthError"
          },
          "422": {
            "$ref": "#/components/responses/AuthError"
          },
          "500": {
            "$ref": "#/components/responses/AuthError"
          }
        }
      }
    },
    "/admin/products/{productId}/publish": {
      "parameters": [
        {
          "$ref": "#/components/parameters/AdminProductId"
        }
      ],
      "post": {
        "operationId": "adminProductsPublish",
        "description": "Publish a valid draft product assigned to an active category.",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Product published.",
            "headers": {
              "Cache-Control": {
                "schema": {
                  "type": "string",
                  "const": "no-store"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminProductResponse"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/AuthError"
          },
          "403": {
            "$ref": "#/components/responses/AuthError"
          },
          "404": {
            "$ref": "#/components/responses/AuthError"
          },
          "409": {
            "$ref": "#/components/responses/AuthError"
          },
          "422": {
            "$ref": "#/components/responses/AuthError"
          },
          "500": {
            "$ref": "#/components/responses/AuthError"
          }
        }
      }
    },
    "/admin/products/{productId}/archive": {
      "parameters": [
        {
          "$ref": "#/components/parameters/AdminProductId"
        }
      ],
      "post": {
        "operationId": "adminProductsArchive",
        "description": "Archive a product without deleting catalog or commercial records.",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Product archived.",
            "headers": {
              "Cache-Control": {
                "schema": {
                  "type": "string",
                  "const": "no-store"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminProductResponse"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/AuthError"
          },
          "403": {
            "$ref": "#/components/responses/AuthError"
          },
          "404": {
            "$ref": "#/components/responses/AuthError"
          },
          "409": {
            "$ref": "#/components/responses/AuthError"
          },
          "422": {
            "$ref": "#/components/responses/AuthError"
          },
          "500": {
            "$ref": "#/components/responses/AuthError"
          }
        }
      }
    },
    "/admin/products/{productId}/images/signature": {
      "parameters": [
        {
          "$ref": "#/components/parameters/AdminProductId"
        }
      ],
      "post": {
        "operationId": "adminProductImagesSignUpload",
        "description": "Create a product-bound, five-minute signed Cloudinary upload contract for JPEG, PNG, or WebP source images up to 4 MB.",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Restricted upload contract.",
            "headers": {
              "Cache-Control": {
                "schema": {
                  "type": "string",
                  "const": "no-store"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminMediaSignatureResponse"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/AuthError"
          },
          "401": {
            "$ref": "#/components/responses/AuthError"
          },
          "403": {
            "$ref": "#/components/responses/AuthError"
          },
          "404": {
            "$ref": "#/components/responses/AuthError"
          },
          "409": {
            "$ref": "#/components/responses/AuthError"
          },
          "413": {
            "$ref": "#/components/responses/AuthError"
          },
          "415": {
            "$ref": "#/components/responses/AuthError"
          },
          "422": {
            "$ref": "#/components/responses/AuthError"
          },
          "500": {
            "$ref": "#/components/responses/AuthError"
          },
          "503": {
            "$ref": "#/components/responses/AuthError"
          }
        }
      }
    },
    "/admin/products/{productId}/images": {
      "parameters": [
        {
          "$ref": "#/components/parameters/AdminProductId"
        }
      ],
      "post": {
        "operationId": "adminProductImagesRegister",
        "description": "Register a signed upload only after the API verifies authoritative provider metadata and product ownership.",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AdminProductImageRegister"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Verified image registered.",
            "headers": {
              "Cache-Control": {
                "schema": {
                  "type": "string",
                  "const": "no-store"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminProductImageResponse"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/AuthError"
          },
          "401": {
            "$ref": "#/components/responses/AuthError"
          },
          "403": {
            "$ref": "#/components/responses/AuthError"
          },
          "404": {
            "$ref": "#/components/responses/AuthError"
          },
          "409": {
            "$ref": "#/components/responses/AuthError"
          },
          "413": {
            "$ref": "#/components/responses/AuthError"
          },
          "415": {
            "$ref": "#/components/responses/AuthError"
          },
          "422": {
            "$ref": "#/components/responses/AuthError"
          },
          "500": {
            "$ref": "#/components/responses/AuthError"
          },
          "502": {
            "$ref": "#/components/responses/AuthError"
          },
          "503": {
            "$ref": "#/components/responses/AuthError"
          }
        }
      }
    },
    "/admin/products/{productId}/images/{imageId}": {
      "parameters": [
        {
          "$ref": "#/components/parameters/AdminProductId"
        },
        {
          "$ref": "#/components/parameters/AdminProductImageId"
        }
      ],
      "patch": {
        "operationId": "adminProductImagesUpdate",
        "description": "Update alternative text, display order, or primary-image selection for a product image.",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AdminProductImageUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Image metadata updated.",
            "headers": {
              "Cache-Control": {
                "schema": {
                  "type": "string",
                  "const": "no-store"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminProductImageResponse"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/AuthError"
          },
          "401": {
            "$ref": "#/components/responses/AuthError"
          },
          "403": {
            "$ref": "#/components/responses/AuthError"
          },
          "404": {
            "$ref": "#/components/responses/AuthError"
          },
          "409": {
            "$ref": "#/components/responses/AuthError"
          },
          "413": {
            "$ref": "#/components/responses/AuthError"
          },
          "415": {
            "$ref": "#/components/responses/AuthError"
          },
          "422": {
            "$ref": "#/components/responses/AuthError"
          },
          "500": {
            "$ref": "#/components/responses/AuthError"
          }
        }
      },
      "delete": {
        "operationId": "adminProductImagesRemove",
        "description": "Remove the image from MartHub and durably track retry-safe provider cleanup. Returns 202 while cleanup remains pending or failed.",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Provider asset cleanup completed.",
            "headers": {
              "Cache-Control": {
                "schema": {
                  "type": "string",
                  "const": "no-store"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminMediaCleanupResponse"
                }
              }
            }
          },
          "202": {
            "description": "Provider asset cleanup remains durable for retry.",
            "headers": {
              "Cache-Control": {
                "schema": {
                  "type": "string",
                  "const": "no-store"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminMediaCleanupResponse"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/AuthError"
          },
          "401": {
            "$ref": "#/components/responses/AuthError"
          },
          "403": {
            "$ref": "#/components/responses/AuthError"
          },
          "404": {
            "$ref": "#/components/responses/AuthError"
          },
          "409": {
            "$ref": "#/components/responses/AuthError"
          },
          "422": {
            "$ref": "#/components/responses/AuthError"
          },
          "500": {
            "$ref": "#/components/responses/AuthError"
          }
        }
      }
    },
    "/admin/products/{productId}/prices": {
      "parameters": [
        {
          "$ref": "#/components/parameters/AdminProductId"
        }
      ],
      "get": {
        "operationId": "adminProductPricesList",
        "description": "List immutable price history newest first, including actor and database-time current state.",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Administrative price history.",
            "headers": {
              "Cache-Control": {
                "schema": {
                  "type": "string",
                  "const": "no-store"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminProductPriceHistoryResponse"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/AuthError"
          },
          "403": {
            "$ref": "#/components/responses/AuthError"
          },
          "404": {
            "$ref": "#/components/responses/AuthError"
          },
          "422": {
            "$ref": "#/components/responses/AuthError"
          },
          "500": {
            "$ref": "#/components/responses/AuthError"
          }
        }
      },
      "post": {
        "operationId": "adminProductPricesCreateSuccessor",
        "description": "Atomically close the latest open predecessor and create its actor-attributed open-ended successor. Middle insertion and timeline rewriting are rejected.",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AdminProductPriceCreate"
              },
              "example": {
                "price": "349000",
                "compareAtPrice": "399000",
                "startsAt": "2026-10-01T00:00:00.000Z"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successor interval created.",
            "headers": {
              "Cache-Control": {
                "schema": {
                  "type": "string",
                  "const": "no-store"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminProductPriceResponse"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/AuthError"
          },
          "401": {
            "$ref": "#/components/responses/AuthError"
          },
          "403": {
            "$ref": "#/components/responses/AuthError"
          },
          "404": {
            "$ref": "#/components/responses/AuthError"
          },
          "409": {
            "$ref": "#/components/responses/AuthError"
          },
          "413": {
            "$ref": "#/components/responses/AuthError"
          },
          "415": {
            "$ref": "#/components/responses/AuthError"
          },
          "422": {
            "$ref": "#/components/responses/AuthError"
          },
          "500": {
            "$ref": "#/components/responses/AuthError"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "AdminSessionResponse": {
        "$schema": "https://json-schema.org/draft/2020-12/schema",
        "type": "object",
        "properties": {
          "data": {
            "type": "object",
            "properties": {
              "user": {
                "type": "object",
                "properties": {
                  "id": {
                    "type": "string",
                    "format": "uuid",
                    "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$"
                  },
                  "email": {
                    "type": "string",
                    "format": "email",
                    "pattern": "^(?:[A-Za-z0-9_'+\\-]+\\.)*[A-Za-z0-9_'+\\-]*[A-Za-z0-9_+-]@(?:[A-Za-z0-9][A-Za-z0-9\\-]*\\.)+[A-Za-z]{2,}$"
                  },
                  "firstName": {
                    "type": "string"
                  },
                  "lastName": {
                    "type": "string"
                  },
                  "role": {
                    "type": "string",
                    "const": "ADMIN"
                  },
                  "phone": {
                    "type": ["string", "null"]
                  },
                  "status": {
                    "type": "string",
                    "enum": ["ACTIVE", "SUSPENDED", "ARCHIVED"]
                  }
                },
                "required": ["id", "email", "firstName", "lastName", "role"],
                "additionalProperties": false
              }
            },
            "required": ["user"],
            "additionalProperties": false
          }
        },
        "required": ["data"],
        "additionalProperties": false
      },
...
    },
...
    "responses": {
...
      "Error": {
        "description": "Normalized API error.",
        "headers": {
          "X-Request-Id": {
            "schema": {
              "type": "string"
            }
          }
        },
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      },
...
    },
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Active session and current database user status/role required."
      },
      "RefreshCookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "mh_refresh",
        "description": "Rotated opaque token. Only SHA-256 hashes are persisted; replay revokes the family."
      }
    }
  }
}